refactor(internal): wrap request
