+++ internal/check.go
@@ -1,0 +1,7 @@
+package internal
+
+type CheckRequest struct {
+	Request
+}
+
+type CheckResponse []Version
